requirements.txt
# Unified Analytics Dependencies
# =============================
# Core dependencies for unified analytics with Plotly integration
# Install with: pip install -r requirements.txt
#
# Compatible with Python 3.8+

# Core Data Processing
# -------------------
pandas>=1.3.0,<3.0.0          # Data manipulation and analysis
numpy>=1.20.0,<2.0.0          # Numerical computing

# Date and Time Handling
# ----------------------
python-dateutil>=2.8.0        # Enhanced date parsing capabilities
pytz>=2021.1                  # Timezone support
PyYAML>=6.0                   # YAML parsing for configuration files

# Data Visualization
# ------------------
matplotlib>=3.5.0,<4.0.0      # Basic plotting and charts (backward compatibility)
seaborn>=0.11.0,<1.0.0        # Statistical data visualization
plotly>=5.0.0,<6.0.0          # Interactive visualizations

# Web Framework
# -------------
Flask>=2.0.0,<3.0.0           # Web framework for UI
Werkzeug>=2.0.0,<3.0.0        # WSGI toolkit for Flask
gunicorn>=20.0.0,<22.0.0      # WSGI HTTP Server
python-dotenv>=0.19.0         # Load environment variables from .env file

# Export Functionality
# -------------------
weasyprint>=60.0              # HTML to PDF conversion (fallback)
reportlab>=3.6.0              # PDF generation with images
selenium>=4.15.0              # Web driver for PNG screenshots
google-api-python-client>=2.0.0  # Google APIs client library
google-auth-httplib2>=0.1.0   # Google Auth HTTP transport
google-auth-oauthlib>=0.5.0   # Google Auth OAuth flow
pillow>=9.0.0                 # Image processing

# Database Connectivity
# --------------------
pymongo>=4.0.0,<5.0.0         # MongoDB driver for Python
dnspython>=2.0.0              # DNS support for MongoDB connection strings

# AI Integration & Enhanced Query Engine
# --------------------------------------
requests>=2.28.0              # HTTP client for API calls
duckdb>=0.8.0                 # In-memory SQL database for AI queries
google-generativeai>=0.3.0   # Google Gemini API client (optional)

# Optional: Enhanced Google Sheets Integration
# (already included above but noting for AI functionality)
# google-api-python-client>=2.0.0
# google-auth-httplib2>=0.1.0
# google-auth-oauthlib>=0.5.0
